Handle non-string values of node.external_id
A change is being proposed to Nodepool to store external_id as
a dictionary.  Prepare the Nodes web page to handle that.

Other than this, the only other part of Zuul that interacts with
external_id is where it is passed to Ansible via the inventory;
that will handle a conversion to a dictionary transparantly.

Change-Id: I2d889c0944981dffdd7abb558c9d92afc1eafd49