Keep project info in repo state keys
In some cases we may produce a repo state after an initial merger
job, then perform a refstate job to add more projects to it.

To decide which projects to add, we consult the repo state from
the first merger job and remove those from the set to request in
the second.  Currently, we load the repo state back into memory
from ZK in order to find out which projects are in it.  We can
make this more efficient if we store the project names in our list
of keys.

Change-Id: I582b2725e66340667339d9ef109500d367b2f81d