Use merger_items to get additional repo state projects
In some cases we may produce a repo state after an initial merger
job, then perform a refstate job to add more projects to it.

To decide which projects to add, we consult the repo state from
the first merger job and remove those from the set to request in
the second.  Currently, we load the repo state back into memory
from ZK in order to find out which projects are in it.  We can
make this more efficient by consulting the merger_items on the
queue item.  This contains a list of all the projects that would
have been included in the original merge (and therefore are
currently in the repo state).

Change-Id: I052aaf86981bbd3d7300c5dcfc374dbad93ab1c1