Add image configuration object
As part of the nodepool-in-zuul work, this adds a stub "image"
configuration object.

It is not (and the following objects will not be) added to the
default include list, meaning that production zuul deployments
(and the bulk of our tests) will not accept these objects.  They
must be enabled explicitly in tenant configuration.  That will
remain the case until we are ready to actually see real-world
use of nodepool-in-zuul functionality.

In the interim, the @simple_layout decorator is updated to allow
tests to opt-in to using these.

Similarly, the documentation will wait until later in the process
since we don't expect this code to be used.

Change-Id: Ic6ee579cb918acb339be510e81d02ca14d65c473