Stop logging refs set
We currently emit a debug log line for each ref that we set when
preparing a repo for a job.  This might be an enormous number of
very boring log lines for large repos.  Since we now store this
information in a file accessible to the job (which can be saved
with the job logs), this information is now accessible for debug
purposes that way.

Instead, just log how many refs we intend to set (so that if there
are a lot, we know what the executor/merger is doing) and log any
missing objects.

Change-Id: Iebc52cbc4aa4fb778b371f90343e53ece6fa3699