Implement auto-reload on new status page
Like the old status page, this will reload the status data every 5
seconds when auto-reload is enabled via the preferences (enabled by
default).

I picked up the idea to use a custom useInterval() react hook from this
article [1]. Using this approach avoids a lot of callback chaining and
takes care about clearing the interval on every re-render.

By using the Page Visibility API in a custom useDocumentVisibility()
hook, we can ensure that the data is only fetched when the page is in
focus. Thus, when the user switches to another tab in the browser or
another window/application, data fetching is skipped.

[1]: https://overreacted.io/making-setinterval-declarative-with-react-hooks/

Change-Id: I39fa6a0a5189665f184aeafcf07ce81365a41902