Add provider objects
This is part of nodepool-in-zuul.  It adds a skeleton provider
object.  These objects behave a little differently than most config
objects.  Like pipelines, they have dual aspects as config definitions
and long-lived objects with real resources attached.

We create model.ProviderConfig objects to represent the config
dictionary.  These perform minimal validation and processing of the
provider.  Once all the nodepool objects have been loaded, we take
the provider config dictionary, merge it with the section hierarchy
and any images/flavors/labels referenced, and take the resulting
flat dictionary and ask the corresponding driver to produce a
real Provider object.  We add that to the layout (if it's a static,
not dynamic, layout).

We will put as many common configuration options as possible in the
BaseProvider class, and add cloud-specific options in the drivers.
This is true both for constructing the actual objects, and also for
creating voluptuous schemas.

This isn't quite the end of the story for providers.  Many of the
resources we will need to track will be based on the Section that
the provider references (for example, our quota values and cloud
api caching should happen on the Section).  A later change will
add a shared section object to the providers created here.

Change-Id: I141cd23428bae548a67012d7382c73781cdf0de1