Add provider Endpoint class
This is part of nodepool-in-zuul.

This class is the closest thing that corresponds to a current
nodepool statemachine.Adapter class.  This holds a real long-lived
connection to a cloud endpoint (that should survive reconfigurations
as long as its config doesn't change).  It will be populated with
methods from the current nodepool statemachine adapter interface.

Later, the launger and builder routines in Zuul will interact with
the Provider objects to decide what images to build and instances
to launch, and the Endpoint object to interact with the cloud to
create those images and launch those nodes.

Change-Id: Ie1b86bd886bdf1d3db5d6fa8d67598accd060bb3