Add ZKObject-based launcher API
This change adds two new ZKObjects, namely the NodesetRequest and a
ProviderNode. A launcher API that operates on cached data provides an
interface that will be used by the launcher server later on.

This change only adds the API and exercises it in a unittest. Following
changes will use it as part of the launcher client/server.

Change-Id: I38528ca75002abca2bdcc7aee2173d0cdcf3129e