Vertically compress the nested list in web job view
The job display shows a list of attributes (like description,
nodeset, file matchers, etc).  Some of these attributes, such
as dependencies, roles, file matchers, have a nested list as
their value.  The leading (vertical spacing) between nested values
is quite large for that purpose.  This makes it difficult to read
the list and takes up a lot of screen space.

Set the margin to 0 for these items (which leaves us with the
normal inter-line spacing).

For an example, see system-config-upload-image-gerrit-3.9 in
the openstack tenant.

Change-Id: I7520fc8acc2623e627f4aa7f64900005db340a88