Restrict section inheritance to a single project
This is part of nodepool-in-zuul.

It may be possible to run the zuul-launcher without a full copy of
the layout in memory.  The schedulers can flatten the provider
configuration and store the result in ZK for the launchers to use.

One aspect that can not be flattened into the provider configuration
is the quota limits that are configured on sections.  It's okay
for multiple providers to have copies of information within a
section object, but we don't intend for every provider to have
its own quota; that should apply to the section itself and the
providers should share that.

That means that we do need to store at least some section configuration
in ZK too, at least the quota limits.  (We might choose to store
more of the section configuration in ZK, and thereby reduce some
duplication in the provider config; that remains to be seen and is not
important for this change).

We can identify a globally unique section using its canonical name,
so we can imagine a ZJ object at:

  /zuul/section/<canonical section name>/config

And providers can reference that section so that the launcher will
know what quota limits to apply to that provider.  And of course
those limits will be shared by any other provider referencing the
section.

Sections can inherit from other sections (so that users can define
a "rax" section and a "rax-dfw" section).  This could cause complications
for the above scheme if two tenants had different "rax" sections but
the same "rax-dfw" section.  We would effectively have two rax-dfw
sections for each tenant (ie, we would require tenant-scoped, not
global-scoped sections).

To avoid this situation, we add the restriction that sections may only
inherit from sections within the same project.  This means that any
section appearing within multiple tenants will be guaranteed to inherit
from the same sections.  Thus, section objects themselves are guaranteed
to be globally unique.

(They may still reference different images/flavors/labels in different
tenants; but that is fine.  The purpose here is to ensure we have a
single configuration for the non-variable parts of sections (eg, quota
limits or any other similar configuration).

Change-Id: I570028d8275734ad1e0e8aa776f227ad2f41da8e