Serialize Provider objects to ZK
This is part of nodepool-in-zuul.

This serializes the provider objects to zookeeper with enough information
for the launcher to deserialize them and create working Providers and
Endpoints without access to an in-memory Layout.  This should allow
the launchers to start and stop quickly and reduce their memory and
CPU requirements.

Change-Id: I5744ff8a1303cbe7286d9959029d914fecbb8446