Handle some TemplateNotFoundErrors
This avoids logging tracebacks in a couple of places where we
might expect to encounter TemplateNotFoundErrors.  This allows us
to recover at a lower level.  In general, we are already cleaning
up a queue item that we can't process; this doesn't change that
(we still won't create user reports) but it does make it tidier.

This removes tracebacks from 6 unit tests.

Change-Id: I3467cdbeadf3cbff2b2a7119fd7f826be9dbdc5b