Add ability to fail tests if tracebacks appear in logs
As we improve exception handling so that errors are recoverable,
we may in the future make changes which cause new exceptions to
be logged in unit tests which, if they don't cause a behavior
change, may go unnoticed.

To prevent that, add an ability to fail unit tests in unexpected
tracebacks show up in logs.

This is not enabled yet, since 84 tests fail with it enabled.
Subsequent changes will fix these incidences and eventually flip
the bit to enable failures.

Change-Id: I349b89db7b1ad46171a2f55fb8e06adcf5d77abc