Mark mqtt test exception as okay
The MQTT tests apparently start the mqtt client and allow it to
fail to connect, and then rely on a monkeypatched publish method
for further testing.  This causes a spurious connection failure
traceback in the logs.

If this happens in production, it's worth seeing the entire
traceback for debugging purposes, so mark this exception as expected
in tests.

This affects 8 unit tests.

Change-Id: I0a9b10f4dc1165b748e572a875dc4d6abbc3c563