Support provider node subclasses in launcher API
Different drivers will need to store additional information about a node
instance in Zookeeper. However, we also want to keep all the nodes in
the same Zookeeper path (/zuul/nodes/<node-uuid>) and with that also in
the same tree cache.

This can be accomplished by making `ProviderNode` a abstract base class
and automatically registering subclasses with the parent class. Each of
those subclasses has a unique "provider id" that is passed as a class
argument.

This ID is stored with the provider node data in Zookeeper and will be
used to select the correct subclass implementation when deserializing
objects.

Change-Id: Ibb337cf7aeb808acc54eb22686d624dcb90db6e5