Support provider node subclasses in launcher API
Different drivers will need to store additional information about a node
instance in Zookeeper. However, we also want to keep all the nodes in
the same Zookeeper path (/zuul/nodes/<node-uuid>) and with that also in
the same tree cache.

This can be accomplished by making `ProviderNode` an abstract base class
and automatically registering subclasses with the parent class.

Since this might be a useful pattern for other objects as well, this is
provided as a separate mixin-class to allow allow variable ZKObjects in
the same ZK path. The actual ZKObject type is encoded as a
zero-terminated byte string before the actual (compressed) data.

Change-Id: Ibb337cf7aeb808acc54eb22686d624dcb90db6e5