Fix spurious secret duplication error
A recent refactoring simplified the detection of duplicate config
objects.  But out of an abundance of caution, secret objects had
a special method used to detect whether they were identical.

This has turned out to be counterproductive since it led to an
error when the normal comparison methods were used.  Since the
only time secrets are compared is in this one instance, let's
correct this by using normal equality checks like other objects.

The main difference between the two methods was that the equality
check verified that the source context was the same.  We consider
that question separately when determining whether a config duplication
is permitted (which is what lead to the supurfluous config errors).

The areDataEqual method is no longer used and is removed.

The test that should have caught this error is corrected.

Change-Id: I79de2f0ad2d9069583838f846814ac4441cd04ac