Fix repo state branches for branch/ref items
With the recent optimization in
I052aaf86981bbd3d7300c5dcfc374dbad93ab1c1 we started using the
merger-items instead of the repo state to decide which projects need to
be included when scheduling the global repo state.

This caused problems for some branch/ref items for which we skip the
initial repo state when unprotected branches are excluded. The
merger-items itself were included in the list of items later on, but we
did not consider the branches for those items.

The effect was that some branches were no longer included in the repo
state.

Change-Id: I0e34a53ab015116759605973ebb72b88c5682c31