Add exclude_results to buildsets api
The web ui shows some boring results (like NO_JOBS and DEQUEUED)
and users probably don't want to see those.  So that we can filter
them out, add an exclude_results option to the buildsets api.

This will mirror what we do for SKIPPED jobs for the builds api.

(For the record, we report the disposition of every buildset that
we enqueue to the database.  Considering that a no-jobs report may
be sent to a remote system and may include a URL pointing to a
buildset, we should store a buildset record in the database.)

Change-Id: I94214cb40adb7b7a06ab02dcb21b3b6d888323bd