Add exclude_results to web UI buildset filtering
Building on the previous change to add exclude_results to the
api, this change adds implicit filtering of buildsets with
results of NO_JOBS or DEQUEUED.

If a user explicitly filters for results, the implicit exclusions
are dropped so that a user may still see a list of DEQUEUED
results if desired.

The feature to drop the implicit exclusion of SKIPPED results was
broken; this change fixes that as well.

Change-Id: I5319edb2546a93e3a62e21a76f330e5f8e81548b