Add lock expiration delays to quick-start
Recent runs of the quick-start job have started to fail and the
apparent cause is that podman may shut down the ZK pod before
the Zuul components have had a chance to release their locks.
This can cause the scheduler to need to wait for lock expiration
before starting the gerrit event stream, which in turn may cause
it to miss important events in the job.

To correct this, add a 30 second delay between stopping and starting
the containers.  When humans run through this sequence, they are
likely to be slow enough not to observe this condition, so we only
need to fix this in the test playbook (not the actual tutorial
instructions).

Change-Id: I806d33237058fa6f0a7c7cc22ceec47242cb4ad1