Add **kw processing to zkobject subclass fromZK
All of the methods of instantiating a ZKObject allow passing
extra kw args to the constructor which will _set attributes on
the object which may not otherwise be serialized.

Add that to the new fromZK method of the variable (polymorphic)
zkobject subclasses.

(Rename one instance of kwargs to kw for consistency.)

Change-Id: Ic36047bfc56210a9979421515af859e2869d2d96