Print stdout on bwrap error
If the early bubblewrap check fails in __init__ it currently throws an
exception and tells the user to run `zuul-bwrap /tmp id` to debug further.
However running that command only leads to the same exception being
thrown without any extra output.
Instead log the output from the command as an error.

Change-Id: Ib4193e7ea9b2fd60d04d810e2f23cdc4994eea41