Don't pre fail jobs if they will retry
Prior to this update it was possible for a job to fail in pre-run which
would result in a retry to then also fail in post-run (possibly because
pre-run failed or due to run not running). In that situation the
post-run Ansible execution would notice early failure which would be
treated by the scheduler as a proper job failure leading to gate resets
and the like. Problem is that once the job finishes and reports it would
then be treated as a retry case leading to inconsistencies in behavior
with queue state and odd reporting to users.

In this situation we don't want to treat a post-run failure as inducing
early failure status. Instead we check the will_retry flag when
processing post-run failures and only report early failure if we won't
retry.

A test is added to cover this which is basically a copy of the existing
pre run failure case but we add in post-run failures.

Change-Id: If78e9b8b55362f5c09d989f6c91a8d773c99fe80