Add image build triggers
This is part of nodepool-in-zuul.

This will trigger image build jobs for all images each time the
layout is updated.  Of course, they should only be triggered if
the image is missing; a later change will store image build results
in ZK and update this code to check that.

In the mean time, this event is only emitted by launchers, and
no one runs launchers yet, so production sites are safe from this
loop.

Change-Id: Id650ef7d8638dfc5e7bd5bd5b6c784a0a6f2b60c