WIP: Try docker compose with podman
DNM: This switches to packaged git-review for simplicity because
pip install does not work in noble; we should either update
the docs to use apt/yum for git-review or switch the docs and
testing to use a venv for git-review.

DNM: Also move catatonit

We want to use podman to run our containers, but the podman-compose
tool has been problematic to use in OpenDev and has also caused
some confusion for users.  It also seems the podman project is
favoring declaring systems using k8s syntax rather than compose,
which is putting the podman-compose tool future in question.

Meanwhile, docker has created a new golang based docker plugin
that is used via the "docker compose" command.  Podman provides
an interface to the docker api, so we can use that to run our
containers via podman (which allows for speculative execution of
containers hosted on any distribution platform) with the docker
compose user interface.

Change-Id: I238f6db178af86fa7203203993c90bd8d1371712
Depends-On: https://review.opendev.org/923083
Depends-On: https://review.opendev.org/925916