Add commentable_files variable to GerritChange
Gerrit will give 400 file not found in revision error on merge changes
with jobs that do gerrit inline commenting. This is due to the fact
that the file list will include the parent 1. The POST API in gerrit
'/changes/{change-id}/revisions/{revision-id}/review' only knows about
the current revision. The filter in the reporter will also fail to
exclude the appropriate files as it includes changes from parent 1.

Change-Id: I64a4879698894f2ef255d769e0cdb59fec64981f
Co-Authored-By: James E. Blair <jim@acmegating.com>