Deprecate cleanup-run
This moves cleanup-run playbooks into the post-run phase to address
a security vulnerability.

To protect against cleanup playbooks being able to access the
resources of playbooks higher in the inheritance stack, we will now
only execute post (including cleanup) playbooks at the level we
reached when running pre-run playbooks.

The cleanup-run attribute is deprecated.  A new attribute to mark
individual post-run playbooks as "cleanup" playbooks is added, and
existing cleanup-run playbooks are treated thusly for backwards
compatibility.

A new deprecation warning is issued for users of cleanup-run.

Change-Id: I49245a4eecb931d8fe7f523805b6e4d7292c806b