Add queue.max-changes
This allows the "Queue" config object (the shared queue definition)
to specify the maximum number of changes associated with that queue
in a given pipeline.

This can be used to protect Zuul from dealing with an excessive number
of changes, but still attempt to localize the fallout to the project(s)
responsible for the overload.

Change-Id: If3de50a74dc8f83fb59f03a0a3fc778cc317d183