Use jaeger all-in-one v1 image
It looks like the :latest image may be the head of the main branch,
which is cool, we do that too, but apparently Jaeger does not test
their releases with Zuul, and it appears they have merged something
that breaks our usage for it.  They may be starting v2 development
and thus may be intentionally merging backwards incompatible changes.

To address this, use the :1 image tag, so that we maintain testing
against the latest v1 jaeger all-in-one image.

Change-Id: I8fef91ac26996b0b10e544006b95bca9381b1262