Show QueueItemPopover on hover
This adds additional event listeners to the Popover and its trigger
element to show/hide it when the cursor is hovering over the trigger
element. It also ensures that the Popover stays open when the cursor
moves from the trigger element to the Popover.

To easier identify to which trigger element a Popover belongs, reduce
the distance between both elements.

The combination of both - the event listeners and the minimal distance -
ensures that the Popover stays open when moving the cursor from the
trigger element to the Popover.

The Popover can still be controlled by clicking on the trigger element.
This is useful if you want to switch to another trigger element, that is
currently overlapped by the active Popover.

Change-Id: Idc66efb7be631fa0c71fcb992152e5bf87934a8b