Fix getting foreign projects during re-enqueue
It seems that due to a bug in the "foreign" project handling during
re-enqueue (after a reconfiguration) that the project of a  change for a
change can be wrongly modified leading to the exception below.

Given the following:
- tenant A:
    - Github project1
    - Gerrit project1
- tenant B:
    - Gerrit project2

From tenant A's perspective Gerrit project2 is a foreign project, but it
is part of tenant B. The issue is now when there is a PR from Github
project1 to a change in Gerrit project2. During re-enqueue Zuul will
create a GITHUB project for GERRIT repo project2 on-the-fly and modify
the project2 change to use the made-up project. It seems like this is
finally responsible for the below fallout in that the wrong connection
is used for updating the Gerrit change in project2.

ERROR zuul.Scheduler: Exception processing pipeline check in tenant TENANT
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 2296, in process_pipelines
    refreshed = self._process_pipeline(
                ^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 2386, in _process_pipeline
    self.process_pipeline_trigger_queue(tenant, tenant_lock, pipeline)
  File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 2630, in process_pipeline_trigger_queue
    self._process_trigger_event(tenant, pipeline, event)
  File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 2663, in _process_trigger_event
    pipeline.manager.addChange(change, event)
  File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 667, in addChange
    self.getDependencyGraph(change, dependency_graph, event,
  File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 891, in getDependencyGraph
    self.getDependencyGraph(needed_change,
  File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 891, in getDependencyGraph
    self.getDependencyGraph(needed_change,
  File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 891, in getDependencyGraph
    self.getDependencyGraph(needed_change,
  [Previous line repeated 16 more times]
  File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 864, in getDependencyGraph
    self.updateCommitDependencies(change, event)
  File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 1065, in updateCommitDependencies
    source.setChangeAttributes(change, **update_attrs)
  File "/opt/zuul/lib/python3.11/site-packages/zuul/driver/github/githubsource.py", line 212, in setChangeAttributes
    return self.connection.updateChangeAttributes(change, **attrs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/zuul/lib/python3.11/site-packages/zuul/connection/__init__.py", line 563, in updateChangeAttributes
    self._change_cache.updateChangeWithRetry(change.cache_stat.key,
  File "/opt/zuul/lib/python3.11/site-packages/zuul/zk/change_cache.py", line 436, in updateChangeWithRetry
    self.set(key, change, version)
  File "/opt/zuul/lib/python3.11/site-packages/zuul/zk/change_cache.py", line 408, in set
    if self._change_cache[key._hash] is not change:
       ~~~~~~~~~~~~~~~~~~^^^^^^^^^^^
KeyError: '2f3303c04b1d93e54e06a87b796550a7c429a799015c31ec559ae342cd864386'

This behavior can be reproduced in a test. In addition there also seem
to be at least one other issue related to foreign projects (also
observable in the test) during job execution:

Traceback (most recent call last):
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/executor/server.py", line 1184, in do_execute
    self._execute()
  File "/home/westphahl/src/opendev/zuul/zuul/tests/base.py", line 662, in _execute
    super()._execute()
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/executor/server.py", line 1420, in _execute
    self.repo_state[task.connection_name][task.project_name]
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^
KeyError: 'gerrit/project2'

This change also fixes the crd reconfig tests that were not using the
provided project arguments.

The fix is to keep using the existing project in case of non-live items
if it can't be found in the current tenant configuration.

Change-Id: I32b5f22be510a3ce80aa224f8893e7a18ddee459