WIP: Demo combined status page
With the new status page, it is not possible to observe the
same change in multiple pipelines simultaneously, nor is it possible
to see different items for the same project in multiple pipelines.

Both of these are useful for continous deployment scenarios, where
a single action (approving a change) may result in several simultanous
pipeline items which also relate to other items from the same or
related projects.

For example in opendev, a change may merge to the system-config
repository that may cause a new image to be deployed and new
documentation to be published in two different pipelines simultaneously,
and jobs in those pipelines may be interacting with (perhaps waiting
for a semaphore, or simply racing!) periodic deployment jobs in
a periodic pipeline.

To address this use case it would be useful to see the actual contents
of multiple pipelines on a single page.  These should be filterable
by change or project.  For the example above, a user may want to
the the contents of all pipelines with a queue item for "system-config".

We may be able to use the pipeline overview page for this if we add
the ability to expand the contents of pipelines from the compact box
view to the full subway-map view.  This change is a partial demonstration
of this (it currently only works for dependent pipelines, and lacks the
filtering described above; these could be added if we like this approach).

This also can make it easier for users to "zoom in" to a pipeline by
allowing them to stay on the same page and expand their areas of interest
(even if those areas span multiple pipelines).

Adding a global toggle for the expansion would allow a global view of
all pipeline contents which may be useful for smaller Zuul installations.

Change-Id: If1648dd4ce8fd0e3fd4c5f191a11c44c78385cfb