Use straight joins in mysql for build/buildset queries
This causes mysql to order the query plan according to the table
order in the query.  For these queries, this is already the best
order.  This should prevent mysql/mariadb from deciding to filter
the enormous ref table with a temporary table, and instead start
the query with an indexed tenant filter.

Change-Id: Ic5cd855d3c8956c1184b7a2dfc19ff1300e4e213