Update foreign project fix to get source by hostname
This modifies the previous fix to the re-enqueue of non-live
changes in foreign projects to always return a new project object
as before, but do so by looking up the source by the hostname of
the old project.

This should address the original issue while maintaining the
characteristic that the re-enqueue process never reuses the Project
objects from the old state.

Change-Id: Ide6cc8075be0f0ff3f49f2ceaedf765beebee385