Use a circular buffer for ansible syntax errors
We previously stored 200 lines of ansible output so that if Ansible
encountered a syntax error, we could include that information in
the build log.  However, it is possible, through the use of include_role,
for Ansible to encounter a syntax error much later than within the
initial 200 lines of output.

To address that, switch to using a circular buffer of the last 200
lines of output so that a recent syntax buffer is always available.

Change-Id: Ice26fe61bb282a9cf9e6a1ce7a6b7855d0495531