Fix race in test_config_update
The test for config updates with polling for the gerrit checks
plugin could race because it may emit an extra change-merged
event due to extra frequent polling in tests.

This could lead to an error shutting down because the merger
may still be busy.

Handle that by waiting for one more polling cycle before
finishing the test.

Change-Id: I28f7ebfae7e2606088bd09608d6f7cf2021b3684