Fix race in test_submitted_together_storm_fast
This test can race if we process gerrit events too slowly.  If we
process the event for change A too slowly, we may end up discovering
change B or C as related changes while processing it.  Our careful
query counting later in the test depends on that not happening
(even though it would be okay in production).

To avoid this, ensure that we have processed the connection event
queue between each step.

Change-Id: Ifa89369caf24bd1c0aa1582820605688fb4f7ab2