Refactor connection configuration
There are some arguments to the connection configuration method
which are used to exclude various types of connectinons since
not all Zuul components need (or should use) all of the connections.

These arguments are a bit haphazard due to the way they were added
over time, and we're adding more connection types that may not
always need to be initialized in the form of providers.

To make it easier to reason about what connections should be
initialized, the existing arguments are replaced with a set of
explicit requests for driver interfaces, all of which default to
False so that each component will state which things it needs.

Change-Id: I197764ec5ca8cccfc6f98ff5f7576d92a9f191fc