Store image build result in ZK
This stores the information about each image build artifact in
ZooKeeper, and consults this when looking for missage images to
build.

The actual ZK path for the image artifact data is slightly different
than what appears in the spec (it is based on a random uuid for
each artifact rather than incorporating the canonical name in
the znode path).  We use a treecache to watch the entire tree
and lookup information about the builds using the cache.

The entries themselves are ZKObjects (ImageBuiltArtifact).  This
makes zk maintenance easy.

Change-Id: Ia5f4df9197ace8bd7eeb7699a2cbfd43168ad24e