Run changed-merged events before ref-updated
When a changed is merged in gerrit, typically a changed-merged
and ref-updated event is generated. If a file contains a change
to job configuration, only the changed-merged event contains
enough information to determine if a reconfigure should occur.
In post pipelines that trigger on ref-updated events, the job
configuration update will often happen after the post job
has started causing it to use the wrong configuration.
Therefore, check to see if we have a matching changed-merged
and ref-updated event and cause the ref-updated event to run
later.

Change-Id: Iba4c32ee3460ab7d60be54bbca093890cd9c86e4