Fix multi-tenant gerrit pre-filters
The Gerrit trigger event pre-filters functionality had a flaw when
multiple tenants used the same connection but did not have the same
pipeline configuration.  Whichever tenant reconfigured most recently
would set the pre-filters for the entire connection based on that
tenant's pipeline configuration.

This feature was based on the Gerrit checks plugin feature, and since
it is deprecated, that feature likely has only ever seen a single user
with a single tenant.  It likely suffers from the same flaw but that
went unnoticed.  This change does not address the checks plugin, but
it does add TODO comments indicating the flaw for clarity.

To correct the issue for pre-filters, we will store the filters for
each tenant separately and then build a superset of filters for all
tenants after reconfiguration.

The filters have hash and eq methods added so that we can avoid storing
duplicate filters for systems with very similar configurations on
each tenant.  This saves processing time during pre-filtering.

Since the driver reconfigure method is not called on a removed tenant,
there is no facility for removing filters of deleted tenants.  This is
a limitation shared not only with the checks plugin support, but also
other drivers such as the timer or zuul drivers.  We should probably
address that in a future change, but it's not critical now since
the current implementation will "fail safe" by simply having more
pre-filters than necessary.  Given that the behavior before this change
was not to have pre-filters at all, that should be fine.  As schedulers
are restarted, excess pre-filters will naturally age out.

Change-Id: Ib879a8983c5444594ee152735dbcfc7f4dba2f80