Refactor ZKObject to harmonize plain and sharded
Now that we have the following axes of zkobjects:
* Sharded/non-sharded
* Polymorphic/non-polymorphic

And both of those are done by sub- or mixin-classes, we are ending
up with a lot of repeated almost-identical code, and it's becoming
difficult to maintain, or add support for the 4th type of object:
Polymorphic-sharded.

To address this, let's lefactor zkobject so that we use the same
load/save methods for sharded/non-sharded, and just replace the
underlying IO class.  We'll use a raw ZK io class for the non-sharded
case, and a sharded io reader/writer for the sharded case.

Part of the complexity of this comes from the fact that currently
the BufferedShard[Reader/Writer] classes also implement compression,
and the polymorphic zkobject needs to read some initial uncompressed
data.

To allow that to happen, we'll separate the concerns and remove
compression from the sharded io class.  That means the ZKObject itself
as a consumer of that class (and also the raw io class) is responsible
for compression.

We have been compressing the data one 1mb znode chunk at a time, and
now we will be compressing the entire data before writing.  That might
briefly use more memory.  However, while we've been doing the same
thing on the reading side, we do ultimately end up with one giant byte
string that we give to json.loads, so we're already incurring that
memory penalty on the read side.  Because of this, we shouldn't see
an overall memory increase.  If we do want to address that, we should
be able to refactor this again to pass in a zlib stream object instead
of the entire compressed data (and likewise the reverse on read).

Note that this may produce fewer ZK operations because by putting the
zlib compression in the writer as we did previously, we ended up
chunking the data based on the *uncompressed* size.  So if we got 4mb
of data, we would chunk it to 1mb, then compress it, which means we
might end up writing only 100k per chunk, using more znodes than
necessary.  Now we will chunk the compressed data and go much closer
to the znode size limit.

Change-Id: I513976a703f035545eaf5bd05a381dbc0c509cda